Issue soft assertion
