+++ src/test/java/edu/hm/hafner/analysis/IssueAssertSoft.java
@@ -1,0 +1,9 @@
+package edu.hm.hafner.analysis;
+
+import org.assertj.core.api.SoftAssertions;
+
+public class IssueAssertSoft extends SoftAssertions{
+    public IssueAssert assertThat(Issue actual) {
+        return proxy(IssueAssert.class, Issue.class, actual);
+    }
+}
